refactor: move study start logic to study session entity
