+++ src/features/study-session-start/index.ts
@@ -1,2 +1,1 @@
-export { startStudy } from "./model/startStudy";
 export { StudySessionStartView } from "./ui/StudySessionStartView";
